# src/main/resources/META-INF/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Students service",
    "description": "Students service",
    "version": "0.1.0"
  },
  "paths": {
    "/groups/new": {
      "post": {
        "tags": ["Группы студентов"],
        "summary": "Добавляет новую группу",
        "requestBody": {
          "description": "Данные новой группы",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/newGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Группа успешно добавлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/newGroupRequest"
                }
              }
            }
          },
          "404": {
            "description": "Наименование группы не уникально",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{id}": {
      "get": {
        "tags": ["Группы студентов"],
        "summary": "Возвращает группу по её идентификатору",
        "description": "Возвращает подробные данные группы поп её id. Если группы с таким id не существует - возвращает ошибку.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Учетные данные корректны",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/newGroupRequest"
                }
              }
            }
          },
          "404": {
            "description": "Не удалось найти группу с указанным id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/groups": {
      "get": {
        "tags": ["Группы студентов"],
        "summary": "Возвращает неколько групп",
        "description": "Возвращает неколько групп студентов отсортирвоанных по id.",
        "parameters": [
          {
            "description": "Принимает целые значения от 0 и выше",
            "name": "pageNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "Принимает целые значения от 1 и выше",
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Выборка из групп студентов отсротированных по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/groupsPageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "errorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "examples": ["yyyy-MM-dd hh:mm:ss.SSS"]
          },
          "httpErrorCode": {
            "type": "number",
            "examples": ["400"]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "newGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Наименования группы. Имя группы должно быть уникальным."
          }
        }
      },
      "groupResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "уникальный идентификатор группы"
          },
          "name": {
            "type": "string",
            "description": "Наименования группы."
          }
        }
      },
      "groupsPageResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "properties": {
              "pageSize": { "type": "number" },
              "pageNumber": { "type": "number" },
              "totalItems": { "type": "number" },
              "totalPages": { "type": "number" },
              "lastPageNumber": { "type": "number" },
              "isEmpty": { "type": "boolean" }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "$ref": "#/components/schemas/groupResponse"
            }
          }
        }
      }
    }
  }
}